# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CrisisMesh Emergency Response API",
    "description": "Gateways 2026 Domain 4: Crisis Command - Multi-Agent Coordination Control Room API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/state": {
      "get": {
        "tags": [
          "State & Plans"
        ],
        "summary": "Get Disaster State",
        "description": "Retrieve full operational snapshot of current disaster response state.",
        "operationId": "get_disaster_state_state_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan/current": {
      "get": {
        "tags": [
          "State & Plans"
        ],
        "summary": "Get Current Plan",
        "description": "Retrieve current proposed or active resource allocation plan.",
        "operationId": "get_current_plan_plan_current_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan/diff": {
      "get": {
        "tags": [
          "State & Plans"
        ],
        "summary": "Get Plan Diff",
        "description": "Retrieve delta between previously approved plan and current revision.",
        "operationId": "get_plan_diff_plan_diff_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan/{plan_id}/approve": {
      "post": {
        "tags": [
          "State & Plans"
        ],
        "summary": "Approve Plan",
        "description": "Authorize resource plan. Strictly enforces Commander role and Ed25519 cryptographic signature.",
        "operationId": "approve_plan_plan__plan_id__approve_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanApproveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/plan/{plan_id}/reject": {
      "post": {
        "tags": [
          "State & Plans"
        ],
        "summary": "Reject Plan",
        "description": "Commander rejects proposed plan with mandatory operational rationale, triggering re-plan.",
        "operationId": "reject_plan_plan__plan_id__reject_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRejectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/scenario/reset": {
      "post": {
        "tags": [
          "Scenario Control"
        ],
        "summary": "Reset Scenario",
        "description": "Reset the operational crisis environment to clean baseline scenario.",
        "operationId": "reset_scenario_scenario_reset_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenario/start": {
      "post": {
        "tags": [
          "Scenario Control"
        ],
        "summary": "Start Scenario",
        "description": "Ingest T0 disaster incidents and trigger multi-agent pipeline.",
        "operationId": "start_scenario_scenario_start_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenario/step": {
      "post": {
        "tags": [
          "Scenario Control"
        ],
        "summary": "Step Scenario",
        "description": "Advance scenario to T+10 change events (submerged car, ambulance breakdown).",
        "operationId": "step_scenario_scenario_step_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StateSnapshot"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/scenario/play": {
      "post": {
        "tags": [
          "Scenario Control"
        ],
        "summary": "Play Scenario",
        "description": "Play through complete disaster response demonstration automatically.",
        "operationId": "play_scenario_scenario_play_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "speed",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 10.0,
              "minimum": 0.1,
              "description": "Playback speed multiplier",
              "default": 1.0,
              "title": "Speed"
            },
            "description": "Playback speed multiplier"
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlanApproveRequest": {
        "properties": {
          "plan_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Hash"
          },
          "signature_hex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signature Hex"
          },
          "nonce": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nonce"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "auto_sign": {
            "type": "boolean",
            "title": "Auto Sign",
            "default": false
          }
        },
        "type": "object",
        "title": "PlanApproveRequest"
      },
      "PlanRejectRequest": {
        "properties": {
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "forbidden_pairs": {
            "anyOf": [
              {
                "items": {
                  "prefixItems": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "string"
                    }
                  ],
                  "type": "array",
                  "maxItems": 2,
                  "minItems": 2
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forbidden Pairs"
          }
        },
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "PlanRejectRequest"
      },
...
      "StateSnapshot": {
        "properties": {
          "incidents": {
            "type": "object",
            "title": "Incidents"
          },
          "units": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Units"
          },
          "hospitals": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Hospitals"
          },
          "verification_labels": {
            "type": "object",
            "title": "Verification Labels"
          },
          "impact_report": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Impact Report"
          },
          "current_plan": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Plan"
          },
          "previous_approved_plan": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Approved Plan"
          },
          "plan_diff": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Diff"
          },
          "commander_briefing": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Commander Briefing"
          },
          "pending_approvals": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Pending Approvals"
          },
          "message_log": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Message Log"
          },
          "negotiation_history": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Negotiation History"
          },
          "active_agents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Active Agents"
          },
          "veto_count": {
            "type": "integer",
            "title": "Veto Count",
            "default": 0
          },
          "step_budget": {
            "type": "integer",
            "title": "Step Budget",
            "default": 20
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "INITIALIZED"
          },
          "dispatch_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatch Error"
          }
        },
        "type": "object",
        "title": "StateSnapshot",
        "description": "Pydantic model representation of the LangGraph state snapshot for APIs & UI."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}